django_erp/requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.20.0
ipdb==0.13.13

# Code quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
pylint==3.0.3
pylint-django==2.5.5

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
